Create basic layout
The layout of the tow main screens have been created:
The first one is the start game screen. Once you click on the text its start an intent which create the pong table.
The second screen is the pong table in whihc you can see the ball and the racquets as well as the divisory line.

Adjustment have been made in the classes to ensure that everything looks as it should